:bug: fix image ratio and display
